add methods

+++ demo/src/main/java/co/edu/udes/activity/backend/demo/dto/EnrollmentRequestDTO.java
@@ -14,8 +14,8 @@ public class EnrollmentRequestDTO {
     private Date enrollmentDate;
 
     private String status;
-    private Long qualification;
-    private Long p1Qualification;
-    private Long p2Qualification;
-    private Long p3Qualification;
+    private Double qualification;
+    private Double p1Qualification;
+    private Double p2Qualification;
+    private Double p3Qualification;
 }
